upodate mining code

+++ install_external_tools.sh
@@ -99,11 +99,13 @@ InstallMosesTools () {
 InstallFastBPE () {
   cd ${tools_ext}
   if [ ! -x fastBPE/fast ] ; then
-    echo " - download fastBPE software from github"
-    wget https://github.com/glample/fastBPE/archive/master.zip
-    unzip master.zip
-    /bin/rm master.zip
-    mv fastBPE-master fastBPE
+    if [ ! -d fastBPE ] ; then
+      echo " - download fastBPE software from github"
+      wget https://github.com/glample/fastBPE/archive/master.zip
+      unzip master.zip
+      /bin/rm master.zip
+      mv fastBPE-master fastBPE
+    fi
     cd fastBPE
     echo " - compiling"
     g++ -std=c++11 -pthread -O3 fastBPE/main.cc -IfastBPE -o fast
